+++ src/frontend/src/components/pages/global/PageLoading/index.tsx
@@ -4,12 +4,10 @@ import * as S from './index.styled'
 import DefaultTemplate from '@/components/templates/DefaultTemplate'
 import { CircularProgress } from '@/components/atoms'
 
-const Page: React.FC = () => (
+export const PageLoading: React.FC = () => (
   <DefaultTemplate>
     <S.Wrapper>
       <CircularProgress />
     </S.Wrapper>
   </DefaultTemplate>
 )
-
-export default Page
